feat(site-redesign): update site title and redirect method
The site title in the `_config.yml` file has been changed from "Лоббизм в Госдуме" to "ДумаБинго". Additionally, the `card.html` layout has been modified to use a meta refresh tag for redirection instead of JavaScript. This change improves the SEO and accessibility of the site by ensuring that the redirection is handled on the client side as soon as the page loads, without the need for executing JavaScript.